(GP-4.1): Identify / Use the Root Tree (#1)
Changed the blob test in the GitTester.java program to have the sample folder (and files) to be created outside of the git directory. Updated the tester to test the genTree() method and genTreesFromIdx() method. Added a line to delete the working_list after the root of the tree was generated in genTreesFromIdx().