(GP-4.2): Link Commit File to Root
Created three new methods: createNewCommit(), getLastCommit(), and updateHead(). createNewCommit() creates a commit file in the objects subdirectory and contains the information of the root tree for this commit, the parent tree (if any), the author of the commit, the date the commit was made, and the commit message. getLastCommit() retrieves the latest documented commit from the HEAD file, and returns an empty string if HEAD is empty. Finally, updateHead updates the HEAD file with the hash of the latest commit file.