(GP-5.4): Checked the commit method (#4)
This commit is a continuation of my previous commits, hence it starting at 5.4 instead of 5.1.

I updated how the date is formatted in the commits, also removed the scanner functionality (it wasn't meant to exist in the first place), and checked to see if everything now functioned correctly, which it did.